requirements/dev.txt
-r base.txt
django-cors-headers

# formatting
ipython
yapf
autopep8
flake8
jedi

## rules
flake8-import-order
flake8-print
flake8-quotes
pep8-naming
flake8-pep3101

